# api/docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.SocialResponse:
    properties:
      education_name:
        type: string
      education_url:
        type: string
      instagram_name:
        type: string
      instagram_url:
        type: string
      linkedin_name:
        type: string
      linkedin_url:
        type: string
      location_name:
        type: string
      location_url:
        type: string
      telegram_name:
        type: string
      telegram_url:
        type: string
      twitter_name:
        type: string
      twitter_url:
        type: string
      user_id:
        type: string
      website_name:
        type: string
      website_url:
        type: string
      youtube_name:
        type: string
      youtube_url:
        type: string
    type: object
  models.SocialUpdate:
    properties:
      education_name:
        type: string
      education_url:
        type: string
      instagram_name:
        type: string
      instagram_url:
        type: string
      linkedin_name:
        type: string
      linkedin_url:
        type: string
      location_name:
        type: string
      location_url:
        type: string
      telegram_name:
        type: string
      telegram_url:
        type: string
      twitter_name:
        type: string
      twitter_url:
        type: string
      user_id:
        type: string
      website_name:
        type: string
      website_url:
        type: string
      youtube_name:
        type: string
      youtube_url:
        type: string
    type: object
  models.TopicCreate:
    properties:
      topic_name:
        type: string
    type: object
  models.TopicResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      level_id:
        type: integer
      level_name:
        type: string
      topic_name:
        type: string
      updated_at:
        type: string
    type: object
  models.TopicUpdate:
    properties:
      id:
        type: integer
      topic_name:
        type: string
    type: object
...
info:
  contact: {}
  title: Asrlan-Monolithic
paths:
...
  /v1/avatarupload:
    post:
      consumes:
      - application/json
      description: File upload
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: File upload
      tags:
      - file-upload
...
  /v1/badgeupload:
    post:
      consumes:
      - application/json
      description: File upload
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: File Name
        in: query
        name: img_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: File upload
      tags:
      - file-upload
...
  /v1/social:
    put:
      consumes:
      - application/json
      description: This Api for updating social
      parameters:
      - description: Update Social Model
        in: body
        name: Social
        required: true
        schema:
          $ref: '#/definitions/models.SocialUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update Social
      tags:
      - social
  /v1/social/{id}:
    get:
      consumes:
      - application/json
      description: This Api for get social
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get Social
      tags:
      - social
...
  /v1/topic:
    post:
      consumes:
      - application/json
      description: This Api for creating a new topic
      parameters:
      - description: TopicCreate Model
        in: body
        name: TopicCreate
        required: true
        schema:
          $ref: '#/definitions/models.TopicCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TopicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Create Topic
      tags:
      - topics
    put:
      consumes:
      - application/json
      description: This Api for updating topic
      parameters:
      - description: Update TopicUpdate Model
        in: body
        name: TopicUpdate
        required: true
        schema:
          $ref: '#/definitions/models.TopicUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TopicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update Topic
      tags:
      - topics
  /v1/topic/{id}:
    delete:
      consumes:
      - application/json
      description: This Api for deleting topic
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete Topic
      tags:
      - topics
    get:
      consumes:
      - application/json
      description: This Api for get topic
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TopicResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get Topic
      tags:
      - topics
  /v1/topics:
    get:
      consumes:
      - application/json
      description: This Api for get all topics
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TopicResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: ListTopics
      tags:
      - topics
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"